Walls with no bottoms and new No Moss Texture
I added the Wall Models without bottom faces to lower the poly count. These Walls have "F" at the end of their names. Example: Corner Wall F is the new model you should be using. I also added a no moss variant for the materials of the walls. In each of the the Wall's folders in Assets > Models > Environment > Materials, there will be a No Moss folder in each of the Walls Folder.